requirements.txt
rich.console